+++ src/component/Modal.js
@@ -1,15 +1,21 @@
 import * as React from 'react';
+import Grid from '@mui/material/Grid';
 import PropTypes from 'prop-types';
 import Button from '@mui/material/Button';
 import { styled } from '@mui/material/styles';
 import Dialog from '@mui/material/Dialog';
 import DialogTitle from '@mui/material/DialogTitle';
 import DialogContent from '@mui/material/DialogContent';
-import DialogActions from '@mui/material/DialogActions';
 import IconButton from '@mui/material/IconButton';
 import CloseIcon from '@mui/icons-material/Close';
 import Typography from '@mui/material/Typography';
-import AddNote from '../component/AddNote';
+import TextField from '@mui/material/TextField';
+import { useState } from 'react';
+import Box from '@mui/material/Box';
+import Stack from '@mui/material/Stack';
+import AddCircleIcon from '@mui/icons-material/AddCircle';
+import CancelIcon from '@mui/icons-material/Cancel';
+import NoteAddIcon from '@mui/icons-material/NoteAdd';
 
 const BootstrapDialog = styled(Dialog)(({ theme }) => ({
   '& .MuiDialogContent-root': {
@@ -49,19 +55,26 @@ BootstrapDialogTitle.propTypes = {
   onClose: PropTypes.func.isRequired,
 };
 
-export default function CustomizedDialogs({children}) {
+export default function CustomizedDialogs() {
   const [open, setOpen] = React.useState(false);
-
+  const [titleNote, setTitleNote] = useState("");
+  const [contentNote, setContentNote] = useState("");
   const handleClickOpen = () => {
     setOpen(true);
   };
   const handleClose = () => {
     setOpen(false);
   };
+  const handleTitleChange = (event) => {
+    setTitleNote(event.target.value);
+  };
+  const handleContentChange = (event) =>{
+    setContentNote(event.target.value);
+  };
 
   return (
-    <div>
-      <Button variant="outlined" onClick={handleClickOpen}>
+    <Grid>
+      <Button sx={{height:60, backgroundColor:'#6ffbff'}} variant="contained" onClick={handleClickOpen} startIcon={<NoteAddIcon />}>
         Add New Note
       </Button>
       <BootstrapDialog
@@ -69,15 +82,50 @@ export default function CustomizedDialogs({children}) {
         aria-labelledby="customized-dialog-title"
         open={open}
       >
-        <BootstrapDialogTitle id="customized-dialog-title" onClose={handleClose}>
-          
-        </BootstrapDialogTitle>
-        <DialogContent dividers>
-          {children}
-        </DialogContent>
-        
-      </BootstrapDialog>
-    </div>
+      <BootstrapDialogTitle id="customized-dialog-title" onClose={handleClose}>
+        Add new note
+      </BootstrapDialogTitle>
+      <DialogContent dividers>
+        <form>  
+          <Typography variant='paragraph' component='p' textAlign='center'>
+            <label>Add a new note subject/title</label> <br />
+          </Typography>
+          <TextField sx = {{marginTop: 1, backgroundColor: '#6ffbff', width: 226}}
+            onChange = {handleTitleChange}
+            required
+            name='title'
+            id="title"
+            label="Required"
+          />
+          <Box sx = {{marginTop: 2}}>
+            <Typography variant='paragraph' component='p' textAlign='center'>
+            <Box>
+              <label>Add content of note</label> <br />
+            </Box>
+          <TextField sx={{marginTop:1, backgroundColor:'#6ffbff', width:226}}
+            required
+            onChange={handleContentChange}
+            name='content'
+            id="content"
+            label="Required"
+            multiline
+            rows={4}
+            inputProps = {{maxLength: 300}}
+          />
+            </Typography><br />
+            </Box>
+          <Stack direction="row" spacing={2}>
+            <Button sx={{width:103, height:40, backgroundColor:'#0097b3'}} variant="contained" onClick={handleClose} disabled={!titleNote||!contentNote} startIcon={<AddCircleIcon />} type='submit' autoFocus>
+              Add
+            </Button>
+            <Button sx={{width:103, height:40, backgroundColor:'#0097b3'}} variant="contained" onClick={handleClose} endIcon={<CancelIcon />}>
+              Cancel
+            </Button>
+          </Stack>
+            </form>
+          </DialogContent>
+          </BootstrapDialog>
+    </Grid>
   );
 }
